requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
influxdb==5.3.0
msgpack==0.6.1
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
